+++ customer_portal/src/epilot/models/components/portalconfig.py
@@ -181,7 +181,7 @@ class PortalConfig:
     contact_identifiers: Optional[List[str]] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contact_identifiers'), 'exclude': lambda f: f is None }})
     r"""Identifiers to identify a contact."""
     contact_secondary_identifier: Optional[str] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('contact_secondary_identifier'), 'exclude': lambda f: f is None }})
-    r"""Secondary identifier to identify a contact other than the email
+    r"""Deprecated. Use contact_identifiers instead.
 
     Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
     """
@@ -223,7 +223,7 @@ class PortalConfig:
     origin: Optional[Origin] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('origin'), 'exclude': lambda f: f is None }})
     r"""Origin of the portal"""
     self_registration: Optional[bool] = dataclasses.field(default=None, metadata={'dataclasses_json': { 'letter_case': utils.get_field_name('self_registration'), 'exclude': lambda f: f is None }})
-    r"""Allow portal user self-registration without a mapped contact
+    r"""Deprecated. Use self_registration_setting instead.
 
     Deprecated field: This will be removed in a future release, please migrate away from it as soon as possible.
     """